Add experiment data download for multiple experiments
If a user is logged in they can download experiment data for regions
they define in a bed file for all (shown) experiments.
